Add more context when failing to unmarshal JSON
As mentioned in #830, we occasionally have a situation where non-JSON
data comes back to a client even though this is not supposed to happen.
Unfortunately, when it does, the error message produced by Go isn't very
good and doesn't help much with debugging.

Here we add a new helper function for unmarshaling JSON in standard or
error responses that behaves normally in the normal cases, and in case
of a failure, adds more context to the standard library's error and
logs.